use conditional compilation to support mac os
Starting in Mac OS Yosemite 10.10, the GNU version of gcc is no longer available.
The gcc command is simply a wrapper around of clang. This change causes
any reference to tr1 header files to abort.

To fix this issue, we use conditional compilation by checking the
__APPLE__ preprocessor macro.

Another change is to specify -arch x86_64 and use -std=gnu89 to compile
GSL code and samtools.